remove func normalize_path to utils/paths
